+++ tests/test_checker/test_noqa.py
@@ -181,6 +181,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS513': 1,
     'WPS514': 1,
     'WPS515': 1,
+    'WPS516': 1,
 
     'WPS600': 1,
     'WPS601': 1,