+++ tests/test_visitors/test_ast/test_functions/test_call_context/test_open_with.py
@@ -87,7 +87,8 @@ def test_open_inside_context_manager(
 ])
 @pytest.mark.parametrize('call', [
     'close()',
-    'some.attr',
+    'open.attr',
+    'open.attr()',
     'obj.open()',
 ])
 def test_regular_functions(